refactor code clean fn
