+++ readme.md
@@ -1,6 +1,16 @@
-<h1>Golden Clover casino</h1>
-<h4>Trabajo Integrador POO - Segundo cuatrimestre</h4>
+<h1>🤞 Golden Clover casino 💰</h1>
+<h3>Trabajo Integrador POO - Segundo cuatrimestre</h3>
 
+<p>En Golden Clover casino se pueden jugar 3 tipos de juegos de azar, además podrás visualizar tus resultados para hacer un seguimiento. A probar suerte! </p>
+
+### Correr en consola
+
+1. Compilar el ts - lo configuramos en /dist
+   ```bash
+   npx tsc
+2. Ejecutar el programa desde la consola: 
+   ```bash
+   node dist/main.js
 <hr>
 
 
@@ -14,15 +24,6 @@
 3. Actualizar repo local en base a remoto:
    ```bash
    git pull origin main
-
-## Correr en consola
-
-1. Compilar el ts - lo configuramos en /dist
-   ```bash
-   npx tsc
-2. Ejecutar el programa desde la consola: 
-   ```bash
-   node dist/main.js
 <h3>Librerías utilizadas:</h3>
 
 - readlineSync
