+++ src/interface/IApuesta.ts
