+++ src/models/Juego.ts
@@ -1,4 +1,4 @@
-//Cualquier juego debe tener: 
+//Siguiendo la idea de GENERALIZACIÓN, cualquier juego debe tener: 
 
 export abstract class Juego {
     //un nombre del tipo string
@@ -12,7 +12,7 @@ export abstract class Juego {
       this.apuestaMinima = apuestaMinima;
     }
     
-    //y el metodo jugar para cualquier clase que herede Juego, todo juego tendra un metodo jugar, con una apuesta number como parametro y devuelve un string, no tiene incidencia en Juego, solo en las clases que lo hereden (en cada logica de cada juego). Obliga a cada clase hija a usar un metodo jugar, se supone que cada juego tiene una manera de jugar puntual
+    //y el metodo jugar para cualquier clase que herede Juego, todo juego, como una INSTANCIA tendra un metodo jugar, con una apuesta number como parametro y devuelve un string, no tiene incidencia en Juego, solo en las clases que lo hereden (en cada logica de cada juego). Obliga a cada clase hija a usar un metodo jugar, se supone que cada juego tiene una manera de jugar puntual
     abstract jugar(apuesta: number): string;
 }
