+++ src/models/Ruleta.ts
@@ -1,5 +1,5 @@
 import { Juego } from "./Juego";
-import { IApuesta } from "../interfaces/IApuesta";
+import { IApuesta } from "../interface/IApuesta";
 import * as readlineSync from "readline-sync";
 
 export class Ruleta extends Juego implements IApuesta {
@@ -25,9 +25,9 @@ export class Ruleta extends Juego implements IApuesta {
     // verificar si ganó, si el numero elegido es igual al ganador
     if (numeroElegido === numeroGanador) {
       const ganancia = apuesta * 3; // gana multiplicado por 3 el monto apostado
-      resultado += `¡Felicitaciones! Tu ganancia es de $${ganancia}. triplicaste tu apuesta`;
+      resultado += `🎊 ¡Felicitaciones! Tu ganancia es de $${ganancia}. triplicaste tu apuesta`;
     } else {
-      resultado += "¡Upss! Perdiste, no salió el número elegido. ¿Volves a jugar?.";
+      resultado += "🤷‍♀️ ¡Upss! Perdiste, no salió el número elegido. ¿Volves a jugar?.";
     }
     return resultado;
   }
